Enum evm_runtime::ExitReason

pub enum ExitReason {
    Succeed(ExitSucceed),
    Error(ExitError),
    Revert(ExitRevert),
    Fatal(ExitFatal),
}

Exit reason.

Variants

Succeed(ExitSucceed)

Machine has succeeded.

Error(ExitError)

Machine returns a normal EVM error.

Revert(ExitRevert)

Machine encountered an explicit revert.

Fatal(ExitFatal)

Machine encountered an error that is not supposed to be normal EVM errors, such as requiring too much memory to execute.

Implementations

impl ExitReason

pub fn is_succeed(&self) -> bool

Whether the exit is succeeded.

pub fn is_error(&self) -> bool

Whether the exit is error.

pub fn is_revert(&self) -> bool

Whether the exit is revert.

pub fn is_fatal(&self) -> bool

Whether the exit is fatal.